+++ CHANGELOG.md
@@ -9,6 +9,8 @@
 ### Changes
 
 * [#2852](https://github.com/clojure-emacs/cider/issues/2852): Convert 1-based column numbers in response map to Emacs' 0-based system.
+* Differentiate between more types in `cider-eldoc`. They used to be just `var` and `fn` and now we have additional handling for
+macros, special forms and methods.
 
 ### Bugs fixed